+++ src/app/core/http/http.service.ts
@@ -1,27 +1,25 @@
 /** Angular Imports */
-import { Inject, Injectable, InjectionToken, Injector, Optional } from '@angular/core';
-import { HttpClient, HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
+import { Inject, Injectable, InjectionToken, Injector, Optional } from "@angular/core";
+import { HttpClient, HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from "@angular/common/http";
 
 /** rxjs Imports */
-import { Observable } from 'rxjs';
+import { Observable } from "rxjs";
 
 /** Custom Interceptors */
-import { ErrorHandlerInterceptor } from './error-handler.interceptor';
-import { CacheInterceptor } from './cache.interceptor';
-import { ApiPrefixInterceptor } from './api-prefix.interceptor';
+import { ErrorHandlerInterceptor } from "./error-handler.interceptor";
+import { CacheInterceptor } from "./cache.interceptor";
+import { ApiPrefixInterceptor } from "./api-prefix.interceptor";
 
 /**
  * HttpClient is declared in a re-exported module, so we have to extend the original module to make it work properly.
  * (see https://github.com/Microsoft/TypeScript/issues/13897)
  */
-declare module '@angular/common/http/http' {
-
+declare module "@angular/common/http/http" {
   /**
    * Augment HttpClient with the added configuration methods from HttpService, to allow in-place replacement of
    * HttpClient with HttpService using dependency injection.
    */
   export interface HttpClient {
-
     /**
      * Enables caching for this request.
      * @param {boolean} forceUpdate Forces request to be made and updates cache entry.
@@ -40,22 +38,18 @@ declare module '@angular/common/http/http' {
      * @returns {HttpClient} The new instance.
      */
     disableApiPrefix(): HttpClient;
-
   }
-
 }
 
 /**
  *  From @angular/common/http/src/interceptor: allows to chain interceptors
  */
 class HttpInterceptorHandler implements HttpHandler {
-
-  constructor(private next: HttpHandler, private interceptor: HttpInterceptor) { }
+  constructor(private next: HttpHandler, private interceptor: HttpInterceptor) {}
 
   handle(request: HttpRequest<any>): Observable<HttpEvent<any>> {
     return this.interceptor.intercept(request, this.next);
   }
-
 }
 
 /**
@@ -66,24 +60,27 @@ class HttpInterceptorHandler implements HttpHandler {
  * For static interceptors that should always be enabled (like ApiPrefixInterceptor), use the standard
  * HTTP_INTERCEPTORS token.
  */
-export const HTTP_DYNAMIC_INTERCEPTORS = new InjectionToken<HttpInterceptor>('HTTP_DYNAMIC_INTERCEPTORS');
+export const HTTP_DYNAMIC_INTERCEPTORS = new InjectionToken<HttpInterceptor>("HTTP_DYNAMIC_INTERCEPTORS");
 
 /**
  * Extends HttpClient with per request configuration using dynamic interceptors.
  */
 @Injectable()
 export class HttpService extends HttpClient {
-
-  constructor(private httpHandler: HttpHandler,
-              private injector: Injector,
-              @Optional() @Inject(HTTP_DYNAMIC_INTERCEPTORS) private interceptors: HttpInterceptor[] = []) {
+  constructor(
+    private httpHandler: HttpHandler,
+    private injector: Injector,
+    @Optional() @Inject(HTTP_DYNAMIC_INTERCEPTORS) private interceptors: HttpInterceptor[] = []
+  ) {
     super(httpHandler);
 
     if (!this.interceptors) {
       // Configure default interceptors that can be disabled here
       this.interceptors = [
         this.injector.get(ApiPrefixInterceptor),
-        this.injector.get(ErrorHandlerInterceptor)
+        this.injector.get(ErrorHandlerInterceptor),
+        //adding caching.
+        this.injector.get(CacheInterceptor),
       ];
     }
   }
@@ -116,16 +113,11 @@ export class HttpService extends HttpClient {
     return new HttpService(
       this.httpHandler,
       this.injector,
-      this.interceptors.filter(i => !(i instanceof interceptorType))
+      this.interceptors.filter((i) => !(i instanceof interceptorType))
     );
   }
 
   private addInterceptor(interceptor: HttpInterceptor): HttpService {
-    return new HttpService(
-      this.httpHandler,
-      this.injector,
-      this.interceptors.concat([interceptor])
-    );
+    return new HttpService(this.httpHandler, this.injector, this.interceptors.concat([interceptor]));
   }
-
 }